# GH-41702: [C++][Parquet] Thrift: generate template method to accelerate reading thrift