… cpp/src/generated/parquet_types.cpp
… cpp/src/generated/parquet_types.h
… cpp/src/generated/parquet_types.tcc