@mapleFU 2024-05-17
@emkornfield @pitrou I've update a patching here. This generated call less virtual functions during deserializing. Would you mind take a look?

I'm not so familiar with thrift compiler, maybe more useful tools can help deserializing