@pitrou 2024-05-17
@mapleFU I didn't know this was possible. This looks neat in the principle. Did you try to run some benchmark?