@mapleFU 2024-05-17
Run in page index: https://github.com/apache/arrow/issues/41702#issuecomment-2116873657

For dense footer it's more useful since readVirt is called for more times