@wgtmac 2024-05-17
I remember there was about 3% speedup reading a sample parquet file.