@pitrou 2024-05-21
Perhaps you can try with the additional benchmarks in https://github.com/apache/arrow/pull/41761