@mapleFU 2024-05-21
On my M1 Pro with Release(O2):

After:

```
Run on (10 X 24.0711 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB
  L1 Instruction 128 KiB
  L2 Unified 4096 KiB (x10)
Load Average: 7.98, 10.79, 8.83
-------------------------------------------------------------------------------------------------------------
Benchmark                                                   Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------------------------
WriteMetadata/num_columns:1/num_row_groups:1            10248 ns        10198 ns        65596 file_size=459 items_per_second=98.0618k/s
WriteMetadata/num_columns:1/num_row_groups:100         708873 ns       701642 ns         1003 file_size=37.383k items_per_second=1.42523k/s
WriteMetadata/num_columns:1/num_row_groups:1000       7027939 ns      7022677 ns           99 file_size=374.885k items_per_second=142.396/s
WriteMetadata/num_columns:10/num_row_groups:1           78750 ns        78709 ns         8900 file_size=3.762k items_per_second=12.705k/s
WriteMetadata/num_columns:10/num_row_groups:100       6751510 ns      6644838 ns          105 file_size=358.835k items_per_second=150.493/s
WriteMetadata/num_columns:10/num_row_groups:1000     67659713 ns     67142800 ns           10 file_size=3.614M items_per_second=14.8936/s
WriteMetadata/num_columns:100/num_row_groups:1         787280 ns       771871 ns          910 file_size=37.352k items_per_second=1.29555k/s
WriteMetadata/num_columns:100/num_row_groups:100     66632500 ns     66540000 ns           10 file_size=3.61693M items_per_second=15.0286/s
WriteMetadata/num_columns:100/num_row_groups:1000   703455917 ns    699385000 ns            1 file_size=36.2887M items_per_second=1.42983/s
WriteMetadata/num_columns:1000/num_row_groups:1       8089713 ns      8087153 ns           85 file_size=376.655k items_per_second=123.653/s
WriteMetadata/num_columns:1000/num_row_groups:100   705972459 ns    702311000 ns            1 file_size=36.4815M items_per_second=1.42387/s
WriteMetadata/num_columns:10000/num_row_groups:1     82793505 ns     82773750 ns            8 file_size=3.82213M items_per_second=12.0811/s
WriteMetadata/num_columns:10000/num_row_groups:100 7789295000 ns   7492551000 ns            1 file_size=369.089M items_per_second=0.133466/s
ReadMetadata/num_columns:1/num_row_groups:1              3022 ns         3021 ns       229889 file_size=459 items_per_second=330.982k/s
ReadMetadata/num_columns:1/num_row_groups:100           59165 ns        59139 ns        11742 file_size=37.383k items_per_second=16.9092k/s
ReadMetadata/num_columns:1/num_row_groups:1000         587111 ns       586972 ns         1189 file_size=374.885k items_per_second=1.70366k/s
ReadMetadata/num_columns:10/num_row_groups:1            13977 ns        13973 ns        50402 file_size=3.762k items_per_second=71.569k/s
ReadMetadata/num_columns:10/num_row_groups:100         475674 ns       475562 ns         1469 file_size=358.835k items_per_second=2.10278k/s
ReadMetadata/num_columns:10/num_row_groups:1000       4743075 ns      4742237 ns          139 file_size=3.614M items_per_second=210.871/s
ReadMetadata/num_columns:100/num_row_groups:1          119355 ns       119308 ns         5747 file_size=37.352k items_per_second=8.38169k/s
ReadMetadata/num_columns:100/num_row_groups:100       5379931 ns      5378835 ns          133 file_size=3.61693M items_per_second=185.914/s
ReadMetadata/num_columns:100/num_row_groups:1000     58173311 ns     58151000 ns           13 file_size=36.2887M items_per_second=17.1966/s
ReadMetadata/num_columns:1000/num_row_groups:1        1285306 ns      1284195 ns          514 file_size=376.655k items_per_second=778.698/s
ReadMetadata/num_columns:1000/num_row_groups:100     59154014 ns     59110667 ns           12 file_size=36.4815M items_per_second=16.9174/s
ReadMetadata/num_columns:10000/num_row_groups:1      15298734 ns     15288065 ns           46 file_size=3.82213M items_per_second=65.4105/s
ReadMetadata/num_columns:10000/num_row_groups:100   597222875 ns    594531000 ns            1 file_size=369.089M items_per_second=1.682/s
```

Before:

```
WriteMetadata/num_columns:1/num_row_groups:1            13997 ns        10952 ns        64411 file_size=459 items_per_second=91.3074k/s
WriteMetadata/num_columns:1/num_row_groups:100        1161928 ns       781421 ns          915 file_size=37.383k items_per_second=1.27972k/s
WriteMetadata/num_columns:1/num_row_groups:1000       9028193 ns      7580868 ns           91 file_size=374.885k items_per_second=131.911/s
WriteMetadata/num_columns:10/num_row_groups:1           87804 ns        81408 ns         8680 file_size=3.762k items_per_second=12.2838k/s
WriteMetadata/num_columns:10/num_row_groups:100       7922727 ns      7032396 ns           96 file_size=358.835k items_per_second=142.199/s
WriteMetadata/num_columns:10/num_row_groups:1000     83557727 ns     72335889 ns            9 file_size=3.614M items_per_second=13.8244/s
WriteMetadata/num_columns:100/num_row_groups:1        1046771 ns       866386 ns          813 file_size=37.352k items_per_second=1.15422k/s
WriteMetadata/num_columns:100/num_row_groups:100     97720995 ns     74290111 ns            9 file_size=3.61693M items_per_second=13.4607/s
WriteMetadata/num_columns:100/num_row_groups:1000  1042585917 ns    773579000 ns            1 file_size=36.2887M items_per_second=1.29269/s
WriteMetadata/num_columns:1000/num_row_groups:1       9320268 ns      8396910 ns           78 file_size=376.655k items_per_second=119.091/s
WriteMetadata/num_columns:1000/num_row_groups:100   789198500 ns    726929000 ns            1 file_size=36.4815M items_per_second=1.37565/s
WriteMetadata/num_columns:10000/num_row_groups:1    105553526 ns     89228125 ns            8 file_size=3.82213M items_per_second=11.2072/s
WriteMetadata/num_columns:10000/num_row_groups:100 9705208125 ns   7941607000 ns            1 file_size=369.089M items_per_second=0.125919/s
ReadMetadata/num_columns:1/num_row_groups:1              3341 ns         3262 ns       215501 file_size=459 items_per_second=306.531k/s
ReadMetadata/num_columns:1/num_row_groups:100           70801 ns        67469 ns        10226 file_size=37.383k items_per_second=14.8215k/s
ReadMetadata/num_columns:1/num_row_groups:1000         697046 ns       661042 ns         1033 file_size=374.885k items_per_second=1.51276k/s
ReadMetadata/num_columns:10/num_row_groups:1            19616 ns        15182 ns        46741 file_size=3.762k items_per_second=65.866k/s
ReadMetadata/num_columns:10/num_row_groups:100         631976 ns       538377 ns         1240 file_size=358.835k items_per_second=1.85743k/s
ReadMetadata/num_columns:10/num_row_groups:1000       5701558 ns      5375484 ns          122 file_size=3.614M items_per_second=186.03/s
ReadMetadata/num_columns:100/num_row_groups:1          137789 ns       128750 ns         5466 file_size=37.352k items_per_second=7.76702k/s
ReadMetadata/num_columns:100/num_row_groups:100       6475114 ns      6090483 ns          118 file_size=3.61693M items_per_second=164.191/s
ReadMetadata/num_columns:100/num_row_groups:1000     64411345 ns     62630000 ns           11 file_size=36.2887M items_per_second=15.9668/s
ReadMetadata/num_columns:1000/num_row_groups:1        1473490 ns      1402757 ns          453 file_size=376.655k items_per_second=712.882/s
ReadMetadata/num_columns:1000/num_row_groups:100     66037220 ns     64025909 ns           11 file_size=36.4815M items_per_second=15.6187/s
ReadMetadata/num_columns:10000/num_row_groups:1      18425749 ns     16564045 ns           44 file_size=3.82213M items_per_second=60.3717/s
ReadMetadata/num_columns:10000/num_row_groups:100   650862958 ns    636789000 ns            1 file_size=369.089M items_per_second=1.57038/s
```